moved gateways to their corresponding usecase
