+++ access_control/grant_access/GrantAccessGateway.go
@@ -1,4 +1,6 @@
-package access_control
+package grant_access
+
+import "github.com/attestify/go-kernel/access_control"
 
 // GrantAccessGateway provides the behavior for persisting a role record
 // Expected Alternative Behaviors
@@ -7,5 +9,5 @@ package access_control
 // Returns one (1) type of error:
 //  - InternalError - Returned if there is an error invoking the implementation of this interface
 type GrantAccessGateway interface {
-	Grant(accessControl AccessControl) error
+	Grant(accessControl access_control.AccessControl) error
 }
